# add vectorized python