@PEZ 2025-01-23
No hand-vectorized solutions accepted here.